requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
numpy
torch
torchvision
Pillow
packaging
pytorch_lightning
